backend/requirements.txt
fastapi==0.111.0
uvicorn==0.29.0
python-multipart==0.0.9
opencv-python-headless==4.9.0.80
ultralytics==8.2.0
numpy==1.26.4
requests==2.31.0
python-dotenv==1.0.1
aiofiles==23.2.1
websockets==12.0
torch==2.3.0
torchvision==0.18.0
Pillow==10.3.0
